multi_agent_system/requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.6.1
python-dotenv>=1.0.0
pydantic>=2.5.3
httpx>=0.26.0
websockets>=12.0
ultralytics>=8.0.222
pillow>=10.2.0
numpy>=1.26.3
pytest>=7.4.4
pytest-asyncio>=0.23.3
